+++ packages/fiori-app-sub-generator/test/int/fiori-elements/expected-output/fpm_v4/webapp/test/integration/opaTests.qunit.js
@@ -1,0 +1,24 @@
+sap.ui.require(
+    [
+        'sap/fe/test/JourneyRunner',
+        'testNameSpace/fpmv4/test/integration/FirstJourney',
+		'testNameSpace/fpmv4/test/integration/pages/TravelMain'
+    ],
+    function(JourneyRunner, opaJourney, TravelMain) {
+        'use strict';
+        var JourneyRunner = new JourneyRunner({
+            // start index.html in web folder
+            launchUrl: sap.ui.require.toUrl('testNameSpace/fpmv4') + '/index.html'
+        });
+
+       
+        JourneyRunner.run(
+            {
+                pages: { 
+					onTheTravelMain: TravelMain
+                }
+            },
+            opaJourney.run
+        );
+    }
+);