+++ packages/btp-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/btp-utils
 
+## 1.0.3
+
+### Patch Changes
+
+-   d638daa: Update @sap/bas-sdk dependency
+
 ## 1.0.2
 
 ### Patch Changes