+++ packages/reload-middleware/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/reload-middleware
 
+## 0.2.21
+
+### Patch Changes
+
+-   Updated dependencies [d638daa]
+    -   @sap-ux/btp-utils@1.0.3
+
 ## 0.2.20
 
 ### Patch Changes